% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce.R
\name{enforceXPtr}
\alias{enforceXPtr}
\title{Enforce a C++ Signature}
\usage{
enforceXPtr(type, args = character())
}
\arguments{
\item{type}{the return type.}

\item{args}{a list of argument types.}
}
\value{
A \code{\link{cppXPtr}} wrapper with signature enforcing.
}
\description{
Wrap a \code{\link{cppXPtr}} call for a given signature (i.e., arguments
and return type). If the user-provided C++ function does not match the
signature, the wrapper throws an informative error.
}
\seealso{
\code{\link{cppXPtr}}
}
